requirements.txt
faiss_cpu
gdown
loguru
numpy
Pillow
scikit_learn
scipy
torch
torchvision
tqdm
